+++ approval-server/pom.xml
@@ -115,6 +115,9 @@
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                    <mainClass>fun.sast.ApprovalApplication</mainClass>
+                </configuration>
             </plugin>
         </plugins>
     </build>
